refactor: 백엔드 프로젝트 구조 변경
sls 디렉토리 내부에 pkg가 위치하도록 변경
