+++ back-end/.gitignore
@@ -26,3 +26,4 @@ vendor
 .DS_Store
 secret.yml
 .env
+node_modules
