+++ back-end/go.mod
@@ -1,0 +1,7 @@
+module abs
+
+go 1.18
+
+require github.com/aws/aws-lambda-go v1.30.0
+
+require github.com/stretchr/testify v1.7.1 // indirect
