+++ back-end/package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "serverless-go-plugin": "^2.2.0"
+  }
+}
