+++ back-end/pkg/slack/history.go
@@ -1,4 +1,4 @@
-package abs
+package slack
 
 import (
 	"fmt"
