+++ back-end/pkg/slack/slack_test.go
@@ -1,4 +1,4 @@
-package abs
+package slack
 
 import (
 	"fmt"
