+++ back-end/serverless.yml
@@ -1,4 +1,4 @@
-service: abs-backend
+service: slack-backend
 frameworkVersion: '3'
 
 provider:
@@ -31,7 +31,7 @@ functions:
     handler: payments/main.go
     events:
       - httpApi:
-          path: /abs/payments
+          path: /slack/payments
           method: get
 #    The following are a few example events you can configure
 #    NOTE: Please make sure to change your handler code to work with those events
