--- back-end/sls/go.mod
@@ -1,10 +1,0 @@
-module github.com/jesh-hub/account-book-slack/back-end
-
-go 1.18
-
-require (
-	github.com/aws/aws-lambda-go v1.30.0
-	github.com/jesh-hub/account-book-slack/back-end/pkg/abs v0.0.0-20220415174156-99a5303702c8
-)
-
-require github.com/cloudflare/ahocorasick v0.0.0-20210425175752-730270c3e184 // indirect
